64-bit fixes: converted intermodule communications to always use 32bit values

+++ src/afterstep/functions.c
@@ -1293,10 +1293,10 @@ send_aswindow_data_iter_func(void *data, void *aux_data)
     ASWindow *asw = (ASWindow *)data ;
 
     SendConfig (module, M_CONFIGURE_WINDOW, asw);
-    SendString (module, M_WINDOW_NAME, asw->w,asw->frame,(unsigned long)asw, asw->hints->names[0], asw->hints->names_encoding[0]);
-    SendString (module, M_ICON_NAME, asw->w, asw->frame, (unsigned long)asw, asw->hints->icon_name, asw->hints->names_encoding[asw->hints->icon_name_idx]);
-    SendString (module, M_RES_CLASS, asw->w, asw->frame, (unsigned long)asw, asw->hints->res_class, asw->hints->names_encoding[asw->hints->res_class_idx]);
-    SendString (module, M_RES_NAME,  asw->w, asw->frame, (unsigned long)asw, asw->hints->res_name, asw->hints->names_encoding[asw->hints->res_name_idx]);
+    SendString (module, M_WINDOW_NAME, asw->w,asw->frame, asw, asw->hints->names[0], asw->hints->names_encoding[0]);
+    SendString (module, M_ICON_NAME, asw->w, asw->frame, asw, asw->hints->icon_name, asw->hints->names_encoding[asw->hints->icon_name_idx]);
+    SendString (module, M_RES_CLASS, asw->w, asw->frame, asw, asw->hints->res_class, asw->hints->names_encoding[asw->hints->res_class_idx]);
+    SendString (module, M_RES_NAME,  asw->w, asw->frame, asw, asw->hints->res_name, asw->hints->names_encoding[asw->hints->res_name_idx]);
     return True;
 }
